refactor(daemon): split module_access client_args into submodules
